fix(src/rooms/get_nb_rooms.c) replace int by size_t in for loop

+++ src/rooms/get_nb_rooms.c
@@ -11,7 +11,7 @@ size_t get_nb_rooms(char ***array3d)
 {
     size_t nb_rooms = 0;
 
-    for (int i = 0; array3d[i]; i++) {
+    for (size_t i = 0; array3d[i]; i++) {
         if (word_array_len(array3d[i]) == 3)
             nb_rooms++;
     }
